requirements.txt
annotated-types==0.7.0
ansiwrap==0.8.4
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asciichartpy==1.5.25
asttokens==2.4.0
async-lru==2.0.4
attrs==22.2.0
autopep8==2.0.4
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.12.2
black==24.10.0
bleach==6.1.0
boolean.py==4.0
camel-ai==0.2.10
certifi==2024.8.30
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
curl_cffi==0.6.2
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
distro==1.9.0
docstring-parser==0.15
entrypoints==0.4
eval_type_backport==0.2.0
exceptiongroup==1.1.3
executing==2.0.1
Farama-Notifications==0.0.4
fastjsonschema==2.18.1
filelock==3.12.4
fonttools==4.43.1
fqdn==1.5.1
fsspec==2024.10.0
gym==0.21.0
gym-notices==0.0.8
gymnasium==0.29.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
identify==2.5.30
idna==3.10
iniconfig==2.0.0
ipykernel==6.6.1
ipython==8.30.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.6
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.4.4
jupyter-events==0.10.0
jupyter-lsp==2.2.2
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.27.3
jupytext==1.15.2
kiwisolver==1.4.5
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.6.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.4.2
nodeenv==1.9.1
notebook==7.2.2
notebook_shim==0.2.3
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
openai==1.56.1
opencv-python==4.10.0.84
overrides==7.4.0
packaging==23.2
pandas==2.2.3
pandoc==2.4
pandocfilters==1.5.0
papermill==2.3.3
parso==0.8.3
pathlib==1.0.1
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==3.11.0
plotly==5.24.1
pluggy==1.5.0
plumbum==1.9.0
ply==3.11
pre-commit==3.5.0
progressbar2==4.5.0
prometheus-client==0.17.1
prompt_toolkit==3.0.47
protobuf==4.25.5
psutil==5.7.2
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.1.0
Pygments==2.16.1
pyparsing==3.1.1
pyrsistent==0.19.3
pytest==8.3.4
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-utils==3.8.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==26.2.0
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.1
ruff==0.5.6
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stable_baselines3==2.3.2
stack-data==0.6.3
sympy==1.13.3
tabulate==0.8.10
tenacity==9.0.0
terminado==0.17.1
textwrap3==0.9.2
tiktoken==0.7.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
torch==2.2.0
tornado==6.4.2
tqdm==4.66.3
traitlets==5.14.3
triton==2.2.0
types-python-dateutil==2.8.19.14
typing_extensions==4.12.2
tzdata==2023.3
uri-template==1.3.0
urllib3==2.2.3
virtualenv==20.24.5
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.9
zipp==3.21.0
